fix: re-wrote implementation and added documentation
