+++ aws_lambda_powertools/utilities/streaming/__init__.py
@@ -1,0 +1,3 @@
+from aws_lambda_powertools.utilities.streaming.s3_object import S3Object
+
+__all__ = ["S3Object"]
