+++ aws_lambda_powertools/utilities/streaming/transformations/__init__.py
@@ -1,6 +1,7 @@
 from aws_lambda_powertools.utilities.streaming.transformations.base import BaseTransform
+from aws_lambda_powertools.utilities.streaming.transformations.csv import CsvTransform
 from aws_lambda_powertools.utilities.streaming.transformations.gzip import GzipTransform
 from aws_lambda_powertools.utilities.streaming.transformations.json import JsonTransform
 from aws_lambda_powertools.utilities.streaming.transformations.zip import ZipTransform
 
-__all__ = ["BaseTransform", "GzipTransform", "JsonTransform", "ZipTransform"]
+__all__ = ["BaseTransform", "GzipTransform", "JsonTransform", "ZipTransform", "CsvTransform"]
