+++ aws_lambda_powertools/utilities/streaming/transformations/zip.py
@@ -1,11 +1,9 @@
-import io
-import typing
+from typing import IO
 from zipfile import ZipFile
 
 from aws_lambda_powertools.utilities.streaming.transformations.base import BaseTransform
 
 
 class ZipTransform(BaseTransform):
-    def transform(self, input_stream: io.RawIOBase) -> ZipFile:
-        input_as_io = typing.cast(typing.IO[bytes], input_stream)
-        return ZipFile(input_as_io, mode="r", **self.kwargs)
+    def transform(self, input_stream: IO[bytes]) -> ZipFile:
+        return ZipFile(input_stream, mode="r", **self.kwargs)
