+++ examples/streaming/src/s3_transform_lzma.py
@@ -1,7 +1,7 @@
 import zipfile
 from typing import Dict
 
-from aws_lambda_powertools.utilities.streaming.s3 import S3Object
+from aws_lambda_powertools.utilities.streaming.s3_object import S3Object
 from aws_lambda_powertools.utilities.streaming.transformations import ZipTransform
 from aws_lambda_powertools.utilities.typing import LambdaContext
